pushing text2time package reorganization
