+++ .gitignore
@@ -3,3 +3,4 @@ pvops/text/__pycache__/
 pvops/text2time/__pycache__/
 examples/*.npy
 docs/_build/
+.pytest_cache/
